refactor: Move bulk of tt2text main back into the direwolf package
Makes cgo wrangling and cross-package stuff much easier!
